+++ src/main/java/chess/dto/ScoreDto.java
@@ -1,8 +1,6 @@
 package chess.dto;
 
 import chess.model.piece.Color;
-import chess.model.util.ScoreResult;
-import java.util.HashMap;
 import java.util.Map;
 
 public class ScoreDto {
